% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configure_year.R
\name{config_year}
\alias{config_year}
\title{configure year Function}
\usage{
config_year(input_year)
}
\arguments{
\item{love}{Do you love cats? Defaults to TRUE.}
}
\description{
This function allows you to express your love of cats.
}
\examples{
config_year()
}
\keyword{config_year}
