% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configue_year.R
\name{configue_year}
\alias{configue_year}
\title{configure year Function}
\usage{
configue_year(input_year)
}
\arguments{
\item{love}{Do you love cats? Defaults to TRUE.}
}
\description{
This function allows you to express your love of cats.
}
\examples{
configue_year()
}
\keyword{configue_year}
