% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_employee.R
\name{employee_number}
\alias{employee_number}
\title{configure year Function}
\usage{
employee_number(data, year)
}
\arguments{
\item{data}{Load your Korea Business Survey data.}

\item{year}{Put \code{year} you want to analyze.}
}
\description{
This function allows you to express your love of cats.
}
\examples{
employee_number()
}
\keyword{employee_number}
