% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IndVariety-package.R
\docType{package}
\name{IndVariety-package}
\alias{IndVariety}
\alias{IndVariety-package}
\title{IndVariety: A package for calculation of industrial variety.}
\description{
Calculation of industrial variety.
}
\author{
\strong{Maintainer}: Kyusik Kim \email{wpkyusik@gmail.com} (0000-0003-3753-3196)

}
\keyword{internal}
