% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_RV.R
\name{RV}
\alias{RV}
\title{Calculation of Related Variety based on Korean Business Survey data.}
\usage{
RV(data, year)
}
\arguments{
\item{data}{Put your Korean Business Survey data dataframe.}

\item{year}{Put \code{year} you want to calculate.}
}
\description{
This function allows you to calculate Related Variety index based on 2015 administration.
}
\examples{
data <- read.table("KoreanBusinessSurvey.txt", sep = "\\t", colClasses = "character")
RV(data = data, year = 2013)
}
