% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcaulte_UV.R
\name{UV}
\alias{UV}
\title{Calculation of Unrelated Variety based on Korean Business Survey data.}
\usage{
UV(data, year)
}
\arguments{
\item{data}{Put your Korean Business Survey data dataframe.}

\item{year}{Put \code{year} you want to calculate.}
}
\description{
This function allows you to calculate Unrelated Variety index based on 2015 administration.
}
\examples{
data <- read.table("KoreanBusinessSurvey.txt", sep = "\\t", colClasses = "character")
UV(data = data, year = 2013)
}
