% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configue_year.R
\name{config_year}
\alias{config_year}
\title{configure year Function}
\usage{
config_year(input_year)
}
\arguments{
\item{input_year}{Put \code{year} you want to analyze.}
}
\description{
This function allows you to express your love of cats.
}
\examples{
config_year()

}
\keyword{config_year}
