% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_employee.R
\name{employee_number}
\alias{employee_number}
\title{Calculation of the number of total workers based on Korean Business Survey data.}
\usage{
employee_number(data, year)
}
\arguments{
\item{data}{Put your Korean Business Survey data dataframe.}

\item{year}{Put \code{year} you want to calculate.}
}
\description{
This function allows you to calculate total workers employed based on 2015 administration.
}
\examples{
data <- read.table("KoreanBusinessSurvey.txt", sep = "\\t", colClasses = "character")
employee_number(data = data, year = 2013)
}
