% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IndVariety-package.R
\docType{package}
\name{IndVariety-package}
\alias{IndVariety}
\alias{IndVariety-package}
\title{IndVariety: A Package for Calculation of Industrial Variety.}
\description{
Calculation of industrial variety.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/kkyusik/IndVariety}
  \item Report bugs at \url{https://github.com/kkyusik/IndVariety/issues}
}

}
\author{
\strong{Maintainer}: Kyusik Kim \email{wpkyusik@gmail.com} (0000-0003-3753-3196)

}
\keyword{internal}
