% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_RV.R
\name{RV}
\alias{RV}
\title{configure year Function}
\usage{
RV(data, year)
}
\arguments{
\item{input_year}{Put \code{year} you want to analyze.}
}
\description{
This function allows you to express your love of cats.
}
\examples{
config_year()
}
\keyword{config_year}
