% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcaulte_UV.R
\name{UV}
\alias{UV}
\title{configure year Function}
\usage{
UV(data, year)
}
\arguments{
\item{data}{Put your Korean Business Survey data.}

\item{input_year}{Put \code{year} you want to analyze.}
}
\description{
This function allows you to express your love of cats.
}
\examples{
UV()
}
\keyword{config_year}
